PR #700 Start moving the contentview up a level Started shifting the contentview class up to libmproxy
