+++ test/test_console_import.py
@@ -1,0 +1,5 @@
+import libmproxy.contentview as cv
+
+
+def test_pass():
+    assert True
